backend/requirements.txt
Python==v3.12.4
Flask==v3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy-serializer-1.4.22